Extended Clustering

+++ Transformation/transformation/src/main/scala/clustering/Row.scala
@@ -3,8 +3,8 @@ package clustering
 /**
   * Created by marc on 19.04.17.
   */
-class Row(var index : Long,
-          var lat : Double,
-          var lon : Double) {
+class Row(var index : Int,
+          var lat : Float,
+          var lon : Float) {
 
 }
